Moving resources impl in resources declaration file
