+++ examples/import/gcp/variables.tf
@@ -3,3 +3,5 @@ variable "gcp_serviceaccount_json" {}
 variable "gcp_network" {}
 variable "gcp_project" {}
 variable "gcp_region" {}
+
+variable "kubeconfig_path" {}
